Refactor the each headersection into its own file (reducers, components and containers)
