+++ src/components/HeaderSections.jsx
@@ -1,5 +1,8 @@
 import React, {Component} from 'react';
-import PropTypes from 'prop-types';
+import HeaderSectionLastBlock from './HeaderSectionLastBlock.jsx';
+import HeaderSectionRPCServer from './HeaderSectionRPCServer.jsx';
+import HeaderSectionTransactions from './HeaderSectionTransactions.jsx';
+import HeaderSectionMiningStatus from './HeaderSectionMiningStatus.jsx';
 
 class HeaderSections extends Component {
   constructor(props){
@@ -7,55 +10,15 @@ class HeaderSections extends Component {
   }
 
   render(){
-    var headerList = this.props.sectionList.map((section, index) => {
-      return (<SectionItem key={index} action={section.action} title={section.title} classId={section.class} icon={section.icon} value={section.value}/>);
-    });
     return(
       <div className="row">
-        {headerList}
+        <HeaderSectionLastBlock />
+        <HeaderSectionRPCServer />
+        <HeaderSectionTransactions />
+        <HeaderSectionMiningStatus />
       </div>
     );
   }
 }
-
-class SectionItem extends Component {
-  constructor(props){
-    super(props);
-  }
-
-  componentDidMount(){
-
-  }
-
-  render(){
-    let imgPath = `../static/images/${this.props.icon}`;
-    return(
-      <div className="col-sm-3">
-        <div className={this.props.classId}>
-          <div className="row">
-            <div className="col-xs-8">
-              <h4>{this.props.title}</h4>
-              {this.props.value}
-            </div>
-            <div className="col-xs-4">
-              <img src={imgPath}/>
-            </div>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-HeaderSections.propTypes = {
-  sectionList: PropTypes.array
-};
-
-SectionItem.propTypes = {
-  classId: PropTypes.string,
-  title: PropTypes.string,
-  icon: PropTypes.string,
-  value: PropTypes.string
-};
 
 export default HeaderSections;
